+++ DataService/src/main/java/com/crafters/DataService/dtos/UpdateItemRequestDTO.java
@@ -14,5 +14,5 @@ public class UpdateItemRequestDTO {
     private String rowType;
     private String collectionName;
     private Map<String,Object> attributes;
-    private Map<String, Integer> yearValue;
+    private Map<String, Double> yearValue;
 }
